moving s3 urn builder to new file to avoid pulling in extra deps
